sem_6/dec/ps6/Cart/requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
bson==0.5.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
ecdsa==0.17.0
fastapi==0.75.0
fonttools==4.31.2
h11==0.13.0
httptools==0.3.0
idna==3.3
kiwisolver==1.4.0
matplotlib==3.5.1
modint==0.4.0
networkx==2.7.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pymongo==4.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
